+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [0.3.0] - 2026-05-27
 
+- [New] Add Twi::Conversation#upload
 - [New] Add Twi.event, Twi.mock.medium
 - [New] Add Twi.conversation, Twi.mock.conversation, Twi.mock.conversation_error
 - [New] Add Twi::Event.params_for